--- components/content/UI/Content.tsx
@@ -1,16 +1,0 @@
-// content 받아와서 넣은 예정
-
-'use client';
-
-import React from 'react';
-import styled from 'styled-components';
-
-const Content = () => {
-  return <ContentContainer>content example</ContentContainer>;
-};
-
-export default Content;
-
-const ContentContainer = styled.section`
-  margin: 8.4rem 0 8rem;
-`;
