+++ package.json
@@ -21,6 +21,7 @@
     "axios": "^1.4.0",
     "eslint": "^8.43.0",
     "eslint-config-next": "13.4.7",
+    "html-react-parser": "^4.0.0",
     "next": "13.4.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
